bin working, example is there
